ORC-47. Fix MacOS compilation warnings.

+++ c++/src/ColumnReader.cc
@@ -380,7 +380,7 @@ namespace orc {
         bits |= readByte() << (i*8);
       }
       float *result = reinterpret_cast<float*>(&bits);
-      return *result;
+      return static_cast<double>(*result);
     }
   };
